+++ src/main/java/api/ApiManager.java
@@ -14,7 +14,7 @@ import io.restassured.specification.RequestSpecification;
 import static io.restassured.RestAssured.given;
 
 /**
- * Manages the creation of RequestSpecification and execution of an ApiRequest entities.
+ * Manages the creation of RequestSpecification and execution of an ApiRequest salesforce.entities.
  */
 public class ApiManager {
