+++ src/main/java/salesforce/config/Header.java
@@ -5,12 +5,10 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package config;
+package salesforce.config;
 
-public class Param {
-    public static final String USERNAME = "username";
-    public static final String PASSWORD = "password";
-    public static final String CLIENT_ID = "client_id";
-    public static final String CLIENT_SECRET = "client_secret";
-    public static final String GRANT_TYPE = "grant_type";
+public class Header {
+    public static final String ACCEPT = "Accept";
+    public static final String CONTENT_TYPE = "Content-Type";
+    public static final String AUTHORIZATION = "Authorization";
 }
