+++ src/main/java/salesforce/config/HeaderValue.java
@@ -5,10 +5,9 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package config;
+package salesforce.config;
 
-public class Header {
-    public static final String ACCEPT = "Accept";
-    public static final String CONTENT_TYPE = "Content-Type";
-    public static final String AUTHORIZATION = "Authorization";
+public class HeaderValue {
+    public static final String APP_JSON = "application/json";
+    public static final String APP_X_FORM = "application/x-www-form-urlencoded";
 }
