+++ src/main/java/salesforce/config/ParamValue.java
@@ -5,9 +5,8 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package config;
+package salesforce.config;
 
-public class HeaderValue {
-    public static final String APP_JSON = "application/json";
-    public static final String APP_X_FORM = "application/x-www-form-urlencoded";
+public class ParamValue {
+    public static final String PASSWORD = "password";
 }
