+++ src/main/java/salesforce/endpointurl/ElementParam.java
@@ -1,4 +1,4 @@
-package endpointurl;
+package salesforce.endpointurl;
 
 public class ElementParam {
     public static final String ID = "Id";
