+++ src/test/java/scenarios/individual/AccountStepsCreate.java
@@ -3,9 +3,9 @@ package scenarios.individual;
 import api.ApiRequestManager;
 import api.ApiResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import endpointurl.Endpoint;
-import entities.Account;
-import entities.CreatedResponse;
+import salesforce.endpointurl.Endpoint;
+import salesforce.entities.Account;
+import salesforce.entities.CreatedResponse;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
