+++ src/test/java/scenarios/individual/AccountStepsGetAll.java
@@ -2,8 +2,8 @@ package scenarios.individual;
 
 import api.ApiRequestManager;
 import api.ApiResponse;
-import endpointurl.Endpoint;
-import entities.CreatedResponse;
+import salesforce.endpointurl.Endpoint;
+import salesforce.entities.CreatedResponse;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
