+++ src/test/java/scenarios/individual/IndividualStepsCreate.java
@@ -3,9 +3,9 @@ package scenarios.individual;
 import api.ApiRequestManager;
 import api.ApiResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import endpointurl.Endpoint;
-import entities.CreatedResponse;
-import entities.Person;
+import salesforce.endpointurl.Endpoint;
+import salesforce.entities.CreatedResponse;
+import salesforce.entities.Person;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
