+++ src/test/java/scenarios/individual/IndividualStepsGet.java
@@ -2,9 +2,9 @@ package scenarios.individual;
 
 import api.ApiRequestManager;
 import api.ApiResponse;
-import endpointurl.ElementParam;
-import endpointurl.Endpoint;
-import entities.CreatedResponse;
+import salesforce.endpointurl.ElementParam;
+import salesforce.endpointurl.Endpoint;
+import salesforce.entities.CreatedResponse;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
