+++ src/test/java/scenarios/individual/IndividualStepsGetAll.java
@@ -2,7 +2,7 @@ package scenarios.individual;
 
 import api.ApiRequestManager;
 import api.ApiResponse;
-import entities.CreatedResponse;
+import salesforce.entities.CreatedResponse;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
