+++ src/test/java/tests/CampaignTest.java
@@ -3,11 +3,10 @@ package tests;
 import api.ApiRequestManager;
 import api.ApiResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import endpointurl.ElementParam;
-import endpointurl.Endpoint;
-import entities.Campaign;
-import entities.CreatedResponse;
-import entities.Product;
+import salesforce.endpointurl.ElementParam;
+import salesforce.endpointurl.Endpoint;
+import salesforce.entities.Campaign;
+import salesforce.entities.CreatedResponse;
 import org.apache.http.HttpStatus;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
