+++ src/test/java/tests/ContactsTest.java
@@ -3,10 +3,10 @@ package tests;
 import api.ApiRequestManager;
 import api.ApiResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import endpointurl.Endpoint;
-import entities.Account;
-import entities.Contact;
-import entities.CreatedResponse;
+import salesforce.endpointurl.Endpoint;
+import salesforce.entities.Account;
+import salesforce.entities.Contact;
+import salesforce.entities.CreatedResponse;
 import org.apache.http.HttpStatus;
 import org.testng.Assert;
 import org.testng.annotations.*;
